Fix: shared objects seem to be garbage collected.
When working with many apks and a local file manager, accesses to any
objects held by Manager created by multiprocessing.Manager create faults
when accessed.
Since the manager should only shut down if the main thread exits or if
its garbage collected, the garbage collection must be the reason as the
main thread is written to run in an infinite loop.
This commit tries to address that by including the
multiprocessing.Manager as a field in the manager class.