Androzoo: Added command line option to repeat queries
This option should only ever be used if the queries provide new apks
each time, e.g. by excluding apks that were already in the results
table, otherwise the apks will be re-analyzed indefinitely.

There might be some apks which will be re-analyzed anyway in case the
number of apks returned from all provided queries is less than twice as
many as the number of workers, which, given the size of the dataset, is
unlikely and a minor overhead.